Allows uses to specify a conversion block that can be used as a default strategy for converting remote to local keys.
For instance, if all local keys are lower camel-cased versions of the underscored remote keys, 
and we have an NSString category that does that conversion, we might specify a block that
passes the remote key into that method and returns it.

This allows a lot of our keys to be converted by default without having to explicitly map them.

Explicitly mapped keys still take precedence over automatically mapped keys.